fix(build): satisfy Rust 1.97 strict clippy

+++ qeli/src/util.rs
@@ -270,7 +270,7 @@ fn broken_down_time(secs: i64, utc: bool) -> (i64, u32, u32, u32, u32, u32) {
         // Howard Hinnant's civil_from_days (proleptic Gregorian).
         let z = days + 719_468;
         let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-        let doe = (z - era * 146_097) as i64;
+        let doe = z - era * 146_097;
         let yoe = (doe - doe / 1460 + doe / 36_524 - doe / 146_096) / 365;
         let y = yoe + era * 400;
         let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
